+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 This project loosely follows [Semantic Versioning](http://semver.org/),
 but is still in early stages of development.
 
+## [0.0.2] - 2016-1127
+
+* Add options
+* Create uglify files
+* Github repo is now named `angular-flapper`
+
 ## [0.0.1] - 2016-1121
 
 * The module is now named `ngFlapper`